backend/pyproject.toml
[project]
name = "device-gateway-ui-backend"
version = "0.1.0"
description = "Backend API for Device Gateway UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.9",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.28.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]